Email Notifications
"This patch provides email notifications, with support for TLS and SMTP
AUTH. It provides the regular notifications on snatch and download. It
provides the ability to give a list of "global" emails, which receive
notifications for all shows and it also allows configuration of email
notification lists on a per show basis."

This is Sluggers pull request from
https://github.com/midgetspy/Sick-Beard/pull/428

I made the changes in the Development branch and changed a few things
to get it to work with the current version of SickBeard.  I've tested
it and it seems to be working great.

Signed-off-by: schumi2004 <schumi2004@gmail.com>